Add a procedural macro to generate mock structs from traits
By annotating a trait with `#[mock_it]`, a mock struct is generated
which implements the trait while using `Mock`s.

This is a first draft of the macro, and only supports traits without
non-self references in the method signatures (no `&str`). The arguments
and return type have to also implement `Default` and the rest of the
traits needed by `Mock` (`Clone`).